Bug fix
The `get_random_train_mask` function used to create a random
train/validation split was replaced by the `RandomTrainMask()` class.
Because Dask dataframe operations are lazily executed, a different mask
was generated each time an instance of `TabularDask` was interacted
with. The `RandomTrainMask()` class ensures that the same random seed is
used for an instance of `TabularDask` to ensure that the
train/validation mask remains consistent. See the tutorial for an usage
example.